refactor: move the cleanup review store into the db crate (#966)
The first commit left the `dedup_suggestions` counts/queue/decide SQL in
`frontend/src/rpc/cleanup.rs`, which would have made it the only module
under `frontend/src/rpc/` issuing SQL outside `#[cfg(test)]`. Move it
down so the data-layer split holds:

- New `db::cleanup::review` (`db/src/cleanup/review.rs`) owns the review
  queue: `review_counts`, `review_queue`, `decide_suggestion`, and
  `REVIEW_QUEUE_MAX`. `decide_suggestion` folds in the validation the
  handler used to do — refuse a `Pending` decision, refuse an unknown or
  already-reviewed row, run the apply primitive before stamping.
- `CleanupStoreError` moves with it, gaining the variants those checks
  need. `Db` / `Payload` / `UnknownToken` are internal faults; the rest
  carry a sentence a caller renders verbatim, and `Apply` wraps
  `CleanupApplyError`. No raw `sqlx::Error` crosses the boundary.
- `db::CleanupPayload` derives `Deserialize`, so the review layer reads
  back the exact JSON the detector wrote instead of a parallel
  `StoredPayload` in the frontend.
- `frontend/src/rpc/cleanup.rs` is now gate-call-map only, with no
  `sqlx` on any path.
- Store tests move to `db/src/cleanup/review/tests.rs` and gain coverage
  for every new `CleanupStoreError` variant; the rpc suite keeps only
  the wrapper's error mapping.
